correction

+++ strip.lua
@@ -577,15 +577,18 @@ local full = false
 local slot=3 --keep torch and fuel
 while slot<17 do
     turtle.select(slot)
-    if turtle.drop() then
-        slot=slot+1
-    else
-        if not full then
-            print("chest is full") --only print this once
-            full = true
+    if turtle.getItemCount()>0 then
+        if turtle.drop() then
+            slot=slot+1
+        else
+            if not full then
+                print("chest is full") --only print this once
+                full = true
+            end
+            os.sleep(30) --wait for 30 seconds
         end
-        os.sleep(30) --wait for 30 seconds
     end
 end
 turtle.select(3)
+turnStripDirection(true)
 turnStripDirection(true)
